requirements.txt
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Assets==0.12
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
pycparser==2.19
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
webassets==0.12.1
Werkzeug==0.16.0
WTForms==2.2.1
